requirements.txt
ray[rllib]==2.5.1
torch==2.0.1
gym==0.26.0
numpy==1.24.3
pandas==2.0.2
PyYAML==6.0
docker==6.1.3
flask==2.3.2
requests==2.31.0
psutil==5.9.5
matplotlib==3.7.1
tqdm==4.65.0
networkx==3.1
scikit-learn==1.2.2
plotly==5.14.1
